refactor: split compact tax advantaged meters
